refactor: move helper functionss to pool tokens utils
