+++ packages/lib/modules/pool/actions/remove-liquidity/RemoveLiquidityProvider.tsx
@@ -15,7 +15,8 @@ import { useRemoveLiquidityPriceImpactQuery } from './queries/useRemoveLiquidity
 import { RemoveLiquidityType } from './remove-liquidity.types'
 import { Address, formatUnits, Hash } from 'viem'
 import { emptyTokenAmounts, toHumanAmountWithAddress } from '../LiquidityActionHelpers'
-import { getPoolActionableTokens, isCowAmmPool } from '../../pool.helpers'
+import { isCowAmmPool } from '../../pool.helpers'
+import { getPoolActionableTokens } from '../../pool-tokens.utils'
 import { isWrappedNativeAsset } from '@repo/lib/modules/tokens/token.helpers'
 import { useRemoveLiquiditySimulationQuery } from './queries/useRemoveLiquiditySimulationQuery'
 import { useRemoveLiquiditySteps } from './useRemoveLiquiditySteps'
