+++ packages/lib/modules/pool/pool-tokens.utils.spec.ts
@@ -15,6 +15,7 @@ import {
   getCompositionTokens,
   getUserReferenceTokens,
   getFlatUserReferenceTokens,
+  getPoolActionableTokens,
 } from './pool-tokens.utils'
 import { ApiToken } from '../tokens/token.types'
 import { PoolToken } from './pool.types'
@@ -74,6 +75,12 @@ function getCompositionTokensURIs(poolExample: PoolExample): (string | null | un
   return getCompositionTokens(pool).map(t => t.logoURI)
 }
 
+function getPoolActionableTokenSymbols(poolExample: PoolExample): string[] {
+  const pool = getApiPoolMock(poolExample)
+
+  return getPoolActionableTokens(pool).map(t => t.symbol)
+}
+
 describe('getDisplayTokens for flat pools', () => {
   it('BAL WETH 80 20', () => {
     expect(getCompositionTokenSymbols(balWeth8020)).toEqual(['BAL', 'WETH'])
@@ -98,6 +105,8 @@ describe('getDisplayTokens for flat pools', () => {
     `)
 
     expect(getFlatUserReferenceTokenSymbols(balWeth8020)).toEqual(['BAL', 'WETH'])
+
+    expect(getPoolActionableTokenSymbols(balWeth8020)).toEqual(['BAL', 'WETH'])
   })
 
   it('osETH Phantom Composable Stable', () => {
@@ -106,6 +115,8 @@ describe('getDisplayTokens for flat pools', () => {
     expect(getUserReferenceTokenSymbols(osETHPhantom)).toEqual(['WETH', 'osETH'])
 
     expect(getFlatUserReferenceTokenSymbols(osETHPhantom)).toEqual(['WETH', 'osETH'])
+
+    expect(getPoolActionableTokenSymbols(osETHPhantom)).toEqual(['WETH', 'osETH'])
   })
 
   it('sDAI weighted', () => {
@@ -114,6 +125,8 @@ describe('getDisplayTokens for flat pools', () => {
     expect(getUserReferenceTokenSymbols(sDAIWeighted)).toEqual(['sDAI', 'wstETH'])
 
     expect(getFlatUserReferenceTokenSymbols(sDAIWeighted)).toEqual(['sDAI', 'wstETH'])
+
+    expect(getPoolActionableTokenSymbols(sDAIWeighted)).toEqual(['wstETH', 'sDAI'])
   })
 
   it('v2 stable with ERC4626 tokens (V2 so no boosted)', () => {
@@ -129,6 +142,11 @@ describe('getDisplayTokens for flat pools', () => {
     expect(getFlatUserReferenceTokenSymbols(v2SepoliaStableWithERC4626)).toEqual([
       'dai-aave',
       'usdc-aave',
+    ])
+
+    expect(getPoolActionableTokenSymbols(v2SepoliaStableWithERC4626)).toEqual([
+      'usdc-aave',
+      'dai-aave',
     ])
   })
 })
@@ -147,6 +165,14 @@ describe('getDisplayTokens for NESTED pools', () => {
       'WXDAI',
       'staBAL3',
     ])
+
+    expect(getPoolActionableTokenSymbols(staBALv2Nested)).toEqual([
+      'USDT',
+      'USDC',
+      'WXDAI',
+      'WETH',
+      'WBTC',
+    ])
   })
 
   it('aura bal (Nested with supportsNestedActions false)', () => {
@@ -155,14 +181,8 @@ describe('getDisplayTokens for NESTED pools', () => {
     expect(getUserReferenceTokenSymbols(auraBal)).toEqual(['B-80BAL-20WETH', 'auraBAL'])
 
     expect(getFlatUserReferenceTokenSymbols(auraBal)).toEqual(['BAL', 'WETH', 'auraBAL'])
-  })
 
-  it('aura bal (Nested with supportsNestedActions false)', () => {
-    expect(getCompositionTokenSymbols(auraBal)).toEqual(['B-80BAL-20WETH', 'auraBAL'])
-
-    expect(getUserReferenceTokenSymbols(auraBal)).toEqual(['B-80BAL-20WETH', 'auraBAL'])
-
-    expect(getFlatUserReferenceTokenSymbols(auraBal)).toEqual(['BAL', 'WETH', 'auraBAL'])
+    expect(getPoolActionableTokenSymbols(auraBal)).toEqual(['B-80BAL-20WETH', 'auraBAL'])
   })
 })
 
@@ -175,6 +195,8 @@ describe('getDisplayTokens for BOOSTED pools', () => {
     expect(getBoostedUnderlyingTokenSymbols(morphoStakeHouse)).toEqual(['USDC', 'wUSDL'])
 
     expect(getFlatUserReferenceTokenSymbols(morphoStakeHouse)).toEqual(['USDC', 'wUSDL'])
+
+    expect(getPoolActionableTokenSymbols(morphoStakeHouse)).toEqual(['USDC', 'wUSDL'])
   })
 
   it('sDAI boosted', () => {
@@ -183,5 +205,7 @@ describe('getDisplayTokens for BOOSTED pools', () => {
     expect(getUserReferenceTokenSymbols(sDAIBoosted)).toEqual(['GNO', 'sDAI'])
 
     expect(getFlatUserReferenceTokenSymbols(sDAIBoosted)).toEqual(['GNO', 'sDAI'])
+
+    expect(getPoolActionableTokenSymbols(sDAIBoosted)).toEqual(['GNO', 'sDAI'])
   })
 })
