+++ packages/lib/modules/pool/pool.helpers.ts
@@ -2,20 +2,18 @@
 import { getChainId, getNetworkConfig } from '@repo/lib/config/app.config'
 import {
   GqlChain,
-  GqlNestedPool,
   GqlPoolBase,
   GqlPoolNestingType,
   GqlPoolStakingGauge,
   GqlPoolStakingOtherGauge,
   GqlPoolTokenDetail,
   GqlPoolType,
-  GqlToken,
   GqlHook,
 } from '@repo/lib/shared/services/api/generated/graphql'
 import { isSameAddress } from '@repo/lib/shared/utils/addresses'
 import { bn } from '@repo/lib/shared/utils/numbers'
 import BigNumber from 'bignumber.js'
-import { isEmpty, isNil, uniqBy } from 'lodash'
+import { isEmpty, isNil } from 'lodash'
 import { Address, getAddress, parseUnits, zeroAddress } from 'viem'
 import { BPT_DECIMALS } from './pool.constants'
 import { isNotMainnet } from '../chains/chain.utils'
@@ -25,12 +23,10 @@ import { getUserTotalBalanceInt } from './user-balance.helpers'
 import { dateToUnixTimestamp } from '@repo/lib/shared/utils/time'
 import { balancerV2VaultAbi } from '../web3/contracts/abi/generated'
 import { supportsNestedActions } from './actions/LiquidityActionHelpers'
-import { getLeafTokens } from '../tokens/token.helpers'
-import { GetTokenFn } from '../tokens/TokensProvider'
 import { vaultV3Abi } from '@balancer/sdk'
-import { TokenCore, PoolListItem, Pool, PoolToken, PoolCore } from './pool.types'
-import { ApiToken } from '../tokens/token.types'
+import { PoolListItem, Pool, PoolCore } from './pool.types'
 import { getBlockExplorerAddressUrl } from '@repo/lib/shared/utils/blockExplorer'
+import { allPoolTokens, isStandardOrUnderlyingRootToken } from './pool-tokens.utils'
 
 /**
  * METHODS
@@ -441,111 +437,6 @@ export function getWarnings(warnings: string[]) {
   return warnings.filter(warning => !isEmpty(warning))
 }
 
-// TODO: refactor into a more generic function that looks for the symbol in any pool token
-export function getActionableTokenSymbol(tokenAddress: Address, pool: Pool): string {
-  const token = getPoolActionableTokens(pool).find(token =>
-    isSameAddress(token.address, tokenAddress)
-  )
-  if (!token) {
-    console.log('Token symbol not found for address ', tokenAddress)
-    return ''
-  }
-
-  return token.symbol
-}
-
-/*
-  Depending on the pool type, iterates pool.poolTokens and returns the list of GqlTokens that can be used in the pool's actions (add/remove/swap).
-
-  For instance:
-    If the pool supports nested actions, returns the leaf tokens.
-    If the pool is boosted, returns the underlying tokens instead of the ERC4626 tokens.
-*/
-export function getPoolActionableTokens(pool: Pool): ApiToken[] {
-  function excludeNestedBptTokens(tokens: ApiToken[]): ApiToken[] {
-    return tokens
-      .filter(token => !isSameAddress(token.address, pool.address)) // Exclude the BPT pool token itself
-      .filter(token => token !== undefined)
-  }
-
-  // TODO add exception for composable pools where we can allow adding
-  // liquidity with nested tokens
-  if (supportsNestedActions(pool)) {
-    return excludeNestedBptTokens(getLeafTokens(pool.poolTokens as PoolToken[]))
-  }
-
-  if (isBoosted(pool)) {
-    return excludeNestedBptTokens(getBoostedGqlTokens(pool))
-  }
-
-  return excludeNestedBptTokens(pool.poolTokens as ApiToken[])
-}
-
-export function getNonBptTokens(pool: Pool) {
-  return pool.poolTokens.filter(token => !token.nestedPool)
-}
-
-export function getNestedBptTokens(poolTokens: PoolToken[]) {
-  return poolTokens.filter(token => token.nestedPool)
-}
-
-// Returns the parent BPT token whose nested tokens include the given child token address
-export function getNestedBptParentToken(poolTokens: PoolToken[], childTokenAddress: Address) {
-  const nestedBptToken = getNestedBptTokens(poolTokens).find(token =>
-    token.nestedPool?.tokens.some(nestedToken =>
-      isSameAddress(nestedToken.address, childTokenAddress)
-    )
-  )
-  if (!nestedBptToken) {
-    throw new Error(
-      `Provided nestedTokenAddress ${childTokenAddress} does not belong to any underlying token amongst the nested pool/s (${getNestedBptTokens(
-        poolTokens
-      )
-        .map(t => t.symbol)
-        .join(' ,')})`
-    )
-  }
-
-  return nestedBptToken
-}
-
-// Returns true if the given token address belongs to a top level standard/underlying token that is not a nestedBpt
-export function isStandardOrUnderlyingRootToken(pool?: Pool, tokenAddress?: Address): boolean {
-  if (!pool || !tokenAddress) return true
-  const token = pool.poolTokens.find(
-    token =>
-      isSameAddress(token.address, tokenAddress) ||
-      isSameAddress(token.underlyingToken?.address || '', tokenAddress)
-  )
-  return token?.hasNestedPool === false
-}
-
-// Returns the top level tokens that is not nestedBpt
-export function getStandardRootTokens(pool: Pool, poolActionableTokens?: ApiToken[]): ApiToken[] {
-  if (!poolActionableTokens) return []
-  return poolActionableTokens.filter(token =>
-    isStandardOrUnderlyingRootToken(pool, token.address as Address)
-  )
-}
-
-// Returns the child tokens (children of a parent nestedBpt)
-export function getChildTokens(pool: Pool, poolActionableTokens?: ApiToken[]): ApiToken[] {
-  if (!poolActionableTokens) return []
-  return poolActionableTokens.filter(
-    token => !isStandardOrUnderlyingRootToken(pool, token.address as Address)
-  )
-}
-
-export function toGqlTokens(
-  poolTokens: PoolToken[],
-  getToken: GetTokenFn,
-  chain: GqlChain
-): GqlToken[] {
-  return poolTokens
-    .map(token => getToken(token.address, chain))
-    .filter((token): token is GqlToken => token !== undefined)
-}
-
 /*
   Allowed pool swaps:
     1. From a standard root token to another standard root token
@@ -562,99 +453,4 @@ export function isPoolSwapAllowed(pool: Pool, token1: Address, token2: Address):
     return false
   }
   return true
-}
-
-/*
-  Returns all the tokens in the structure of the given pool:
-  top level tokens + children nested tokens + ERC4626 underlying tokens.
- */
-export function allPoolTokens(pool: Pool | GqlPoolBase): TokenCore[] {
-  const extractUnderlyingTokens = (token: PoolToken): TokenCore[] => {
-    if (shouldUseUnderlyingToken(token, pool)) {
-      return [{ ...token.underlyingToken, index: token.index } as TokenCore]
-    }
-    return []
-  }
-
-  const extractNestedUnderlyingTokens = (nestedPool?: GqlNestedPool): TokenCore[] => {
-    if (!nestedPool) return []
-    return nestedPool.tokens.flatMap(nestedToken =>
-      shouldUseUnderlyingToken(nestedToken as PoolToken, pool)
-        ? ([
-            nestedToken,
-            { ...nestedToken.underlyingToken, index: nestedToken.index },
-          ] as TokenCore[])
-        : [nestedToken as TokenCore]
-    )
-  }
-
-  const poolTokens: PoolToken[] = pool.poolTokens as PoolToken[]
-
-  const underlyingTokens: TokenCore[] = poolTokens.flatMap(extractUnderlyingTokens)
-
-  const nestedParentTokens: PoolToken[] = poolTokens.flatMap(token =>
-    token.nestedPool ? token : []
-  )
-
-  const nestedChildrenTokens: TokenCore[] = pool.poolTokens.flatMap(token =>
-    token.nestedPool ? extractNestedUnderlyingTokens(token.nestedPool as GqlNestedPool) : []
-  )
-
-  const isTopLevelToken = (token: PoolToken): boolean => {
-    if (token.hasNestedPool) return false
-    if (!isV3Pool(pool)) return true
-    if (!token.isErc4626) return true
-    if (token.isErc4626 && !token.isBufferAllowed) return true
-    return true
-  }
-
-  const standardTopLevelTokens: PoolToken[] = poolTokens.flatMap(token =>
-    isTopLevelToken(token) ? token : []
-  )
-
-  const allTokens = underlyingTokens.concat(
-    toTokenCores(nestedParentTokens),
-    nestedChildrenTokens,
-    toTokenCores(standardTopLevelTokens)
-  )
-
-  // Remove duplicates as phantom BPTs can be both in the top level and inside nested pools
-  return uniqBy(allTokens, 'address')
-}
-
-function shouldUseUnderlyingToken(token: PoolToken, pool: Pool | GqlPoolBase): boolean {
-  if (isV3Pool(pool) && token.isErc4626 && token.isBufferAllowed && !token.underlyingToken) {
-    // This should never happen unless the API some some inconsistency
-    throw new Error(
-      `Underlying token is missing for ERC4626 token with address ${token.address} in chain ${pool.chain}`
-    )
-  }
-  // Only v3 pools should underlying tokens
-  return isV3Pool(pool) && token.isErc4626 && token.isBufferAllowed && !!token.underlyingToken
-}
-
-// Returns top level standard tokens + Erc4626 (only v3) underlying tokens
-export function getBoostedGqlTokens(pool: Pool): ApiToken[] {
-  const poolTokens = pool.poolTokens as PoolToken[]
-  const underlyingTokens = poolTokens
-    .flatMap(token =>
-      shouldUseUnderlyingToken(token, pool)
-        ? [{ ...token, ...token.underlyingToken } as ApiToken]
-        : [token as ApiToken]
-    )
-    .filter((token): token is ApiToken => token !== undefined)
-  return underlyingTokens
-}
-
-function toTokenCores(poolTokens: PoolToken[]): TokenCore[] {
-  return poolTokens.map(
-    t =>
-      ({
-        address: t.address as Address,
-        name: t.name,
-        symbol: t.symbol,
-        decimals: t.decimals,
-        index: t.index,
-      }) as TokenCore
-  )
 }
